[ADD] Shift mixin
